requirements.txt
transformers
python-dotenv
torch